GradientM_Arena-main/requirements.txt
Flask==3.1.1
openai==1.95.1
azure-identity==1.23.0
python-dotenv==1.1.1
gunicorn